pulls throwError, normaliseValue, and validateProps out of utilities and puts them in the component file
